fixed the manager
I somehow missed the fact that in order for the manager to start with the venv correctly, it needs to be given the venv path. I forgor to do such. Because of this, the code would not have worked at all. And as a result it needed to be recommitted. Am I stalling for time? No! Am I trying to make the message really long? YEP! Anyway I made a tiny change to the line that allows for it to find the venv, which before was wrong on just about all levels lol. So now it should work alongside the rest of the code until I need to make another commit with an even longer message to convey the ideas that I have just presented here.